requirements.txt
streamlit
requests
python-dotenv
pandas
numpy
Pillow
altair
